requirements.txt
PyJWT
requests
flask
gunicorn
redis
